Initial specs migration. (#793)
Fixes #792

An initial migration of the existing specs as of
[`e59efd63a2165866584833e91e1cb8a6ed8c8203`](https://github.com/celestiaorg/celestia-specs/tree/e59efd63a2165866584833e91e1cb8a6ed8c8203).
It's expected that there are a few broken links and extra content that
has to be pruned, along with incorrect content. The important thing is
to not miss anything.

Also sets up a CI workflow that will build the book.