refactor: remove duplciate `validateHash` (#1077)
## Description
It looks like the helper function `validateHash` was intentionally
duplicated in the `da` package to avoid import issues when this package
lived in celestia-core. Now that `da` lives in celestia-app, it seems
possible to import celestia-core
[`ValidateHash`](https://github.com/celestiaorg/celestia-core/blob/2ec23f804691afc196d0104616e6c880d4c1ca41/types/validation.go#L30-L41)
directly rather than duplicate it.


## Motivation
Observed while reading `da` package to answer an unrelated question.