fix: remove inaccurate `BlobSharesUsed` (#1147)
Follow up to https://github.com/celestiaorg/celestia-app/pull/1137 that
fixes Godocs and removes an inaccurate helper function.

While investigating
https://github.com/informalsystems/audit-celestia/issues/14 I noticed
that the `BlobSharesUsed` helper still exists. It should be removed
because it assumes blobs are prefixed with varints (no longer the case)
and doesn't differentiate between the first and continuation shares in a
sparse share sequence. The helper `SparseSharesNeeded` should be used
instead.