feat! implement token filter IBC middleware (#1219)
Closes: https://github.com/celestiaorg/celestia-app/issues/235

This PR creates and wires up a new IBC middleware that acts as a
firewall, rejecting all `FungibleTokenPacketData` (i.e. transfer
packets) that have a denom which did not originally came from the
celestia chain.

This simple implemenation will mean that the chain state only consists
of the native celestia token.