feat!: remove token voting upgrades (#1629)
## Overview

wraps the sdk's standard upgrade module to remove the ability to submit
upgrade proposals. The idea being that this maintains compatibility with
the IBC module (which requires access to the upgrade module's keeper),
while still removing the ability for the upgrade module to work by not
registering the upgrade module's msg server or begin block. Also, when
we implement #1014 we can progressively flesh this module out with our
own logic.

closes #1571 

## Checklist

- [x] New and updated code has appropriate documentation
- [x] New and updated code has new and/or updated testing
- [x] Required CI checks are passing
- [x] Visual proof for any user facing features like CLI or
documentation updates
- [x] Linked issues closed with keywords

---------

Co-authored-by: Rootul P <rootulp@gmail.com>