feat: create compact share counter (#1657)
As part of ADR 020, we need something to iteratively calculate how many
shares both pfbs and txs would take up should they be added. This PR
creates a `CompactShareCounter` to perform this job. It would have been
also possible to extend the `CompactShareSplitter` to do this task but
we would have needed to add the functionality to "undo" transactions
(because they were too large) and this seemed like quite some
functionality to add.

Co-authored-by: Rootul P <rootulp@gmail.com>