chore: update `TestGetDataCommitment` table tests comment (#1762)
## Overview

Adds a simple comment to the `TestGetDataCommitment` not to run the
table tests separately. The reason for this is that the nature of the
test is sequential. However, since it was confusing, we went for a table
structure so it's easy to read. But that doesn't change its sequential
nature.

I tried using a map structure to iterate over the tests which would
still have the table structure, but would execute them sequentially, but
it gets complicated to run them in order, or printing the test names in
case they fail.