doc: emphesizes the exact range of reserved namespaces (#2137)
## Overview

Upon reviewing the namespace module, it came to my attention that
namespaces below the [maximum reserved namespace
](https://github.com/celestiaorg/celestia-app/blob/5615205760ac3921506e2c457fc2deb924f8aed5/pkg/namespace/consts.go#L56)are
being treated as reserved (see
[IsReserved](https://github.com/celestiaorg/celestia-app/blob/2b74a9d970618de76f82d385f46f176452c12bb3/pkg/namespace/namespace.go#L133)),
regardless of their current usage. Therefore, this pull request aims to
revise the specifications to accurately reflect this fact. It is
important to note that the list of reserved namespaces extends beyond
what is included in the table and encompasses namespaces below the
maximum reserved namespace.

## Checklist

- [ ] New and updated code has appropriate documentation
- [ ] New and updated code has new and/or updated testing
- [ ] Required CI checks are passing
- [ ] Visual proof for any user facing features like CLI or
documentation updates
- [ ] Linked issues closed with keywords