test: tests tx inclusion in the first block (#2220)
## Overview
As part of #1899, we need to test the network/validator's behavior for
transaction inclusion in the first block (more precisely, when the first
block has not been committed yet). This PR introduces a new test that
utilizes a local consensus config option, namely `timeout_commit`, to
slow down block production. This delay allows transactions to arrive at
the time of block height 1.

It is important to note that the current test demonstrates that
transaction inclusion fails due to the absence of commit info for block
height 0. There are two theories regarding this behavior:

1) Theory 1: This is expected behavior as the initial state or genesis
state might not be available by the protocol definition. Consequently,
the first block remains empty because there is no valid state available
to validate transactions against. (Our fork of Cosmos SDK is not in
parity with the mainline Cosmos SDK, however, with the initial
inspection, it seems the initial state does not get committed in the
mainline Cosmos SDK too, hence it is suspected that the same problem
exists there as well. However, further tests should be developed to
confirm this)

2) Theory 2: This is not an expected behavior. The genesis state should
be committed when initializing the chain (in `abci.InitChain`). In this
case, the root cause should be found and the fix should be implemented.

In either case, subsequent PRs will be addressing this issue.

## Checklist

- [x] New and updated code has appropriate documentation
- [x] New and updated code has new and/or updated testing
- [x] Required CI checks are passing
- [x] Visual proof for any user facing features like CLI or
documentation updates
- [x] Linked issues closed with keywords