test: genesis state tests for vesting accounts (#2004)
closes #1806 
<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.
-->

Overview

This PR is proposing a test suit to test the vesting accounts where
introduced at genesis state. It tests the scenarios that we are intend
to use for our mainnet.

**Note:** THIS PR is a copy of the PR #1993 as having the refactors of
the main rebased into it to have easier review.

The suggested changes are applied to the tests.
Thank you @cmwaters and @evan-forbes and special thanks to @rootulp for
his detailed points and for his great proposal to make the tests more
fine grains and more special.

<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 
-->

## Checklist

<!-- 
Please complete the checklist to ensure that the PR is ready to be
reviewed.

IMPORTANT:
PRs should be left in Draft until the below checklist is completed.
-->

- [ ] New and updated code has appropriate documentation
- [ ] New and updated code has new and/or updated testing
- [ ] Required CI checks are passing
- [ ] Visual proof for any user facing features like CLI or
documentation updates
- [ ] Linked issues closed with keywords

---------

Co-authored-by: Rootul P <rootulp@gmail.com>
Co-authored-by: CHAMI Rachid <chamirachid1@gmail.com>