chore(deps): bump gogo proto and ethereum version (#2338)
<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.
-->

## Overview

<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 
-->

Now that gogo proto also cut a new release using the latest `x/exp`
package, we can bump its version and also the
version of Ethereum.

This would help me fix dependencies downstream in orchestrator-relayer
and testground tests.

We can also bump its version in cosmos-sdk, but I am not sure whether we
want to do that ourselves or wait until we get that from upstream. There
is a note in gogoproto stating:

> Dependencies depending on earlier version of x/exp may not be able to
bump to v1.4.11 because slices.SortFunc had some breaking changes. For
limiting breaking changes in patch versions, for instance, the Cosmos
SDK v0.47.x will not bump to v1.4.11.


## Checklist

<!-- 
Please complete the checklist to ensure that the PR is ready to be
reviewed.

IMPORTANT:
PRs should be left in Draft until the below checklist is completed.
-->

- [ ] New and updated code has appropriate documentation
- [ ] New and updated code has new and/or updated testing
- [ ] Required CI checks are passing
- [ ] Visual proof for any user facing features like CLI or
documentation updates
- [ ] Linked issues closed with keywords