chore: default home directory for `single-node.sh` (#2431)
Closes https://github.com/celestiaorg/celestia-app/issues/2404

I frequently need to modify a config and test it out. When using the
existing single-node.sh script, it's a bit tedious to find the randomly
generated CELESTIA_APP_HOME directory and plumb that into subsequent
ad-hoc commands.

This PR modifies the `single-node.sh` script to use the default
CELESTIA_APP_HOME directory (usually `~/.celestia-appd`). That way, if a
developer needs to make a modification to a config, they can do that in
the expected location (i.e. `~/.celestia-appd/config/config.toml`) and
then re-run `celestia-appd start` and test out the modification.

Other improvements include:
- Adds a y/n prompt to confirm deleting any existing CELESTIA_APP_HOME
directory
- Hoists variables to the top
- Silences noisy terminal output by redirecting to dev/null

In addition to reviewing the code diff, it may make sense for reviewers
to try the script and see if it behaves as expected.