Makefile: simpily ledger build tags (#2449)
In the process of working on the `goreleaser` PR I discovered that how
we are building the `BUILD_TAGS` is overly verbose and unnecessary.

Ultimately we are just trying to add `ledger` to `-tags`. Hardcoding it
and adding a note in the README for the 2 distributions that aren't
supported seemed easier to maintain.