feat!: allow for hardcoded upgrade schedules (#2583)
This PR implements the protocol work behind ADR018. After several
iterations I have come across a design which I think best meets our
requirements. I will need to update the ADR afterwards.

This still relies on a hardcoded upgrade schedule but rather than simply
upgrading at that height through `EndBlocker`, a proposer will modify
the proposed block of the height before with a message indicating a
version change. Nodes will vote on that version change in
`ProcessProposal`. If 2/3+ support that app version then the proposal
will pass and the block be committed. In the following height the
upgraded nodes will perform the state migration if any and propose a
block corresponding to that new app version. Nodes that don't support
that app version will panic in `DeliverTx`.

---------

Co-authored-by: Rootul P <rootulp@gmail.com>