refactor!: move all blob share commitment code to the inclusion package (#2770)
## Overview

We currently have code relating to creating and proving commitment
across the inclusion, x/blob/types, and shares package. This api
breaking refactor moves all of the commitment code to the inclusion
package.

The reasoning behind this is that it will eventually allow for users to
import commitment creation and proving code while not having to import
the blob module, tendermint, or the sdk.

The reasoning behind keeping it in the inclusion instead of shares is
that entire reason for making a commitment is to prove inclusion. So the
name seems to fit better.

The reasoning behind note keeping it in the blob package was that it was
difficult to not have an import cycle since this logic requires the
share splitting logic for blobs.

## Checklist

- [x] New and updated code has appropriate documentation
- [x] New and updated code has new and/or updated testing
- [x] Required CI checks are passing
- [x] Visual proof for any user facing features like CLI or
documentation updates
- [ ] Linked issues closed with keywords