feat!: add a signalling mechanism for coordinated upgrades (#2832)
As per ADR-018, this PR extends the existing minimal upgrade module with
a signalling mechanism.

Validators are expected to submit an on-chain message to signal that
they wish to change version of the network. When a quorum has signalled
the next version the upgrade module signals to the app that it is ready
to switch to the next state machine. If the app version is not supported
the node will panic. Note that this feature does not currently support
downgrading. The only permissible app version change is the very next
increment. To cancel the upgrade, the same validators need only to
submit an on-chain message with the current version they are on.

Co-authored-by: Evan Forbes <42654277+evan-forbes@users.noreply.github.com>
Co-authored-by: Rootul P <rootulp@gmail.com>
Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>